requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
peewee==3.17.8
psycopg2==2.9.10
pycparser==2.22
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
